Moved PathFinderController class to move_to_pose
